+++ packages/adminui-vue/components/UbToolbarComponent.vue
@@ -155,7 +155,7 @@ module.exports = {
 
               this.$UB.Repository('ubm_form').attrs(['ID', 'code'])
                 .where('code', '=', this.formCode).select()
-                .done(result => {
+                .then(result => {
                   if (!result || result.length < 1) {
                     this.$notify({
                       title: this.$ut('formNotFound'),