+++ packages/cs-shared/CustomRepository.js
@@ -10,7 +10,7 @@ const _ = require('lodash')
 const bracketsRe = /\[.*]/
 // in case values for where is null we transform condition to allowed null comparison with warning.
 // If condition not in conditionInCaseValueIsNull object keys we raise error
-const conditionInCaseValueIsNull = {equal: 'isNull', notEqual: 'notIsNull', custom: 'custom'}
+const conditionInCaseValueIsNull = { equal: 'isNull', notEqual: 'notIsNull', custom: 'custom' }
 
 const cNames = []
 for (let i = 0; i < 100; i++) {
@@ -130,20 +130,42 @@ class CustomRepository {
  // Get attribute value for multilaguage ("isMultiLang": true in meta file) attribute other when current session language
  UB.Repository('org_employee').attrs(['ID', 'lastName_en^']).selectAsObject()
 
-   * @param {string|Array<string>} attr
+   * @param {string|Array<string>} attrs
    * @return {CustomRepository}
    */
-  attrs (attr) {
-    const L = arguments.length
+  attrs (...attrs) {
+    const L = attrs.length
     for (let i = 0; i < L; i++) {
-      let attrI = arguments[i]
-      if (Array.isArray(attrI)) {
-        this.fieldList = this.fieldList.concat(attrI)
+      if (Array.isArray(attrs[i])) {
+        this.fieldList = this.fieldList.concat(attrs[i])
       } else {
-        this.fieldList.push(attrI)
+        this.fieldList.push(attrs[i])
       }
     }
     return this
+  }
+
+  /**
+   * Helper method for {@link class:CustomRepository#attrs CustomRepository.attrs}.
+   * Calls `attrs` in case addingCondition is <a href=https://developer.mozilla.org/en-US/docs/Glossary/truthy>truthy</a>
+   *
+   * @example
+
+   let isPessimisticLock = !!UB.connection.domain.get('uba_user').attributes.mi_modifyDate
+   // with whereIf
+   let repo = UB.Repository('uba_user').attrs('ID').attrsIf(isPessimisticLock, 'mi_modifyDate')
+   //without whereIf
+   let repo = UB.Repository('uba_user').attrs('ID')
+   if (isPessimisticLock) repo = repo.attrs('mi_modifyDate')
+
+   * @param {*} addingCondition Attributes will be added only in case addingCondition is truthy
+   * @param {string|Array<string>} attrs
+   * @return {CustomRepository}attrs
+   */
+  attrsIf (addingCondition, ...attrs) {
+    return addingCondition
+      ? this.attrs.apply(this, attrs)
+      : this
   }
 
   /**
@@ -181,7 +203,9 @@ UB.Repository('my_entity').attrs('id')
    * @param {CustomRepository.WhereCondition|String} condition  Any value from {@link CustomRepository#WhereCondition WhereCondition}
    * @param {*} [values] Condition value. If `undefined` values not passed to ubql
    * @param {string} [clauseName] Optional clause name to be used in {CustomRepository.logicalPredicates}
-   *   If not passed where will generate unique clause named 'c1', 'c2', ......
+   *   If not passed unique clause name will be generated ('_1', '_2', ..).
+   *   In case a condition with the same name exists, it will be overwritten.
+   *
    * @return {CustomRepository}
    */
   where (expression, condition, values, clauseName) {
@@ -215,7 +239,7 @@ UB.Repository('my_entity').attrs('id')
       console.warn('Condition "in" is passed to CustomRepository.where but values is null or undefined -> condition transformed to (0=1). Check your logic')
       expression = '0'
       condition = WhereCondition.equal
-      values = {a: 1}
+      values = { a: 1 }
     } else if (condition === 'in' && (!Array.isArray(values))) {
       console.debug('Condition "in" is passed to CustomRepository.where but values is not an array -> condition transformed to equal. Check your logic')
       condition = WhereCondition.equal
@@ -223,12 +247,12 @@ UB.Repository('my_entity').attrs('id')
       console.warn('Condition "in" is passed to CustomRepository.where but value is empty array -> condition transformed to "0=1". Check your logic')
       expression = '0'
       condition = WhereCondition.equal
-      values = {a: 1}
+      values = { a: 1 }
     } else if (condition === 'notIn' && (!values || !values.length)) {
       console.warn('Condition "notIn" is passed to CustomRepository.where but value is empty array -> condition transformed to "1=1". Check your logic')
       expression = '1'
       condition = WhereCondition.equal
-      values = {a: 1}
+      values = { a: 1 }
     } else if (values === null && (condition !== 'isNull' || condition !== 'notIsNull')) {
       let wrongCondition = condition
       values = undefined
@@ -261,6 +285,34 @@ UB.Repository('my_entity').attrs('id')
     }
     this.whereList[clauseName] = whereItem
     return this
+  }
+
+  /**
+   * Helper method for {@link class:CustomRepository#where CustomRepository.where}.
+   * Calls `where` in case addingCondition is <a href=https://developer.mozilla.org/en-US/docs/Glossary/truthy>truthy</a>
+   *
+   * @example
+
+   let filterString = 'foundAllLikeThis' // or may be empty string
+   // with whereIf
+   let repo = UB.Repository('my_entity').attrs('ID').whereIf(filterString, 'myAttr', 'like', filterString)
+
+   //without whereIf
+   let repo = UB.Repository('my_entity').attrs('ID')
+   if (filterString) repo = repo.where('myAttr', 'like', filterString)
+
+   * @param {*} addingCondition Where expression will be added only in case addingCondition is truthy
+   * @param {string} expression   Attribute name (with or without []) or valid expression with attributes in []
+   * @param {CustomRepository.WhereCondition|String} condition  Any value from {@link CustomRepository#WhereCondition WhereCondition}
+   * @param {*} [values] Condition value. If `undefined` values not passed to ubql
+   * @param {string} [clauseName] Optional clause name to be used in {CustomRepository.logicalPredicates}
+   *   If not passed where will generate unique clause named 'c1', 'c2', ......
+   * @return {CustomRepository}
+   */
+  whereIf (addingCondition, expression, condition, values, clauseName) {
+    return addingCondition
+      ? this.where(expression, condition, values, clauseName)
+      : this
   }
 
   /**
@@ -382,6 +434,7 @@ UB.Repository('tst_document').attrs(['ID', '[caregory.code]'])
     return this
   }
 
+  // noinspection JSUnusedGlobalSymbols
   /**
    * Adds join condition. Fix some known issues
    *
@@ -540,6 +593,7 @@ inst.run('select', repo.ubql())
     return req
   }
 
+  // noinspection JSMethodCanBeStatic
   /**
    * Must be implemented in descendants and return (or resolved for async clients)
    * to `array of object` representation of result, like this
@@ -554,6 +608,7 @@ inst.run('select', repo.ubql())
   selectAsObject (fieldAliases) {
     throw new Error('abstract')
   }
+  // noinspection JSMethodCanBeStatic
   /**
    * Must be implemented in descendants and return (or resolved for async clients)
    * to `array of array` representation of result, like this
@@ -565,7 +620,7 @@ inst.run('select', repo.ubql())
   selectAsArray () {
     throw new Error('abstract')
   }
-
+  // noinspection JSMethodCanBeStatic
   /**
    * Must be implemented in descendants and return (or resolved for async clients)
    * to `DataSet` class instance, implemented in caller level. It can be:
@@ -581,7 +636,7 @@ inst.run('select', repo.ubql())
   selectAsStore (storeConfig) {
     throw new Error('abstract')
   }
-
+  // noinspection JSMethodCanBeStatic
   /**
    * Must be implemented in descendants as a alias to the most appropriate method
    *
@@ -592,6 +647,7 @@ inst.run('select', repo.ubql())
     throw new Error('abstract')
   }
 
+  // noinspection JSMethodCanBeStatic
   /**
    * Select a single row. If ubql result is empty - return `undefined`
    * @example
@@ -608,6 +664,7 @@ inst.run('select', repo.ubql())
     throw new Error('abstract')
   }
 
+  // noinspection JSMethodCanBeStatic
   /**
    * Execute select and returns a value of the first attribute from the first row
    * @example
@@ -625,6 +682,7 @@ inst.run('select', repo.ubql())
     throw new Error('abstract')
   }
 
+  // noinspection JSMethodCanBeStatic
   /**
    * Select a single row by ID. If result is empty - returns `undefined`
    * If result is not empty - returns a object
@@ -667,6 +725,20 @@ inst.run('select', repo.ubql())
     return this
   }
 
+  /**
+   * Helper method for {@link class:CustomRepository#misc CustomRepository.misc}.
+   * Calls `misc` in case addingCondition is <a href=https://developer.mozilla.org/en-US/docs/Glossary/truthy>truthy</a>
+   *
+   * @param {*} addingCondition flags will be applied only in case addingCondition is truthy
+   * @param {Object} flags
+   * @return {CustomRepository}
+   */
+  miscIf (addingCondition, flags) {
+    _.assign(this.__misc, flags)
+    return this
+  }
+
+  // noinspection JSUnusedGlobalSymbols
   /**
    * Calculate total row number. WARNING!! This is VERY slow operation on DB level in case of many record
    *